[jb] add track for sent closed heartbeat
